third_party/WebKit/LayoutTests/fast/mediarecorder/MediaRecorder-requestData.html

<!DOCTYPE html>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<script>

var test = async_test('checks requestData() for video-only MediaRecorder API.');
var recorder;

recorderOnDataAvailable = test.step_func(function(event) {
    assert_equals(event.data.size, 0, 'Recorded data size should be == 0');
    test.done();
});

recorderOnStop = test.step_func(function() {
    assert_unreached('Recording stopped.');
});

recorderOnError = test.step_func(function() {
    assert_unreached('Recording error.');
});

gotStream = test.step_func(function(stream) {
    assert_equals(stream.getAudioTracks().length, 0);
    assert_equals(stream.getVideoTracks().length, 1);
    assert_equals(stream.getVideoTracks()[0].readyState, 'live');

    try {
        recorder = new MediaRecorder(stream);
    } catch (e) {
        assert_unreached('Exception while creating MediaRecorder: ' + e);
    }
    assert_equals(recorder.state, "inactive");

    try {
        recorder.requestData();
        assert_unreached();
    } catch (e) {
        assert_equals(e.name, "InvalidStateError");
    }

    recorder.ondataavailable = recorderOnDataAvailable;
    recorder.onstop = recorderOnStop;
    recorder.onerror = recorderOnError;
    recorder.start();

    assert_equals(recorder.state, "recording");
    recorder.requestData();
});

onError = test.step_func(function() {
    assert_unreached('Error creating MediaRecorder.');
});

try {
    navigator.webkitGetUserMedia({video:true}, gotStream, onError);
} catch(e) {
    assert_unreached('Exception launching getUserMedia(): ' + e);
}

</script>
